+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/AudioFile.kt
@@ -5,6 +5,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
 import com.github.insanusmokrassar.TelegramBotAPI.types.FileUniqueId
 import com.github.insanusmokrassar.TelegramBotAPI.types.fileUniqueIdField
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.utils.MimeType
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -18,7 +19,7 @@ data class AudioFile(
     override val performer: String? = null,
     override val title: String? = null,
     @SerialName(mimeTypeField)
-    override val mimeType: String? = null,
+    override val mimeType: MimeType? = null,
     @SerialName(fileSizeField)
     override val fileSize: Long? = null,
     override val thumb: PhotoSize? = null