+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/DocumentFile.kt
@@ -4,6 +4,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
 import com.github.insanusmokrassar.TelegramBotAPI.types.FileUniqueId
 import com.github.insanusmokrassar.TelegramBotAPI.types.fileUniqueIdField
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.utils.MimeType
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -17,7 +18,7 @@ data class DocumentFile(
     override val fileSize: Long? = null,
     override val thumb: PhotoSize? = null,
     @SerialName(mimeTypeField)
-    override val mimeType: String? = null,
+    override val mimeType: MimeType? = null,
     @SerialName(fileNameField)
     override val fileName: String? = null
 ) : TelegramMediaFile, MimedMediaFile, ThumbedMediaFile, CustomNamedMediaFile