+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/VideoFile.kt
@@ -4,6 +4,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
 import com.github.insanusmokrassar.TelegramBotAPI.types.FileUniqueId
 import com.github.insanusmokrassar.TelegramBotAPI.types.fileUniqueIdField
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.utils.MimeType
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -18,7 +19,7 @@ data class VideoFile(
     override val duration: Long? = null,
     override val thumb: PhotoSize? = null,
     @SerialName(mimeTypeField)
-    override val mimeType: String? = null,
+    override val mimeType: MimeType? = null,
     @SerialName(fileSizeField)
     override val fileSize: Long? = null
 ) : TelegramMediaFile, MimedMediaFile, ThumbedMediaFile, PlayableMediaFile, SizedMediaFile