+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/VoiceFile.kt
@@ -1,9 +1,9 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.files
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
-import com.github.insanusmokrassar.TelegramBotAPI.types.FileUniqueId
-import com.github.insanusmokrassar.TelegramBotAPI.types.fileUniqueIdField
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.utils.MimeType
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -13,9 +13,10 @@ data class VoiceFile(
     override val fileId: FileId,
     @SerialName(fileUniqueIdField)
     override val fileUniqueId: FileUniqueId,
+    @SerialName(durationField)
     override val duration: Long? = null,
     @SerialName(mimeTypeField)
-    override val mimeType: String? = null,
+    override val mimeType: MimeType? = null,
     @SerialName(fileSizeField)
     override val fileSize: Long? = null
 ) : TelegramMediaFile, MimedMediaFile, PlayableMediaFile