+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/MediaGroupMemberInputMediaSerializer.kt
@@ -1,6 +1,9 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia
 
+import com.github.insanusmokrassar.TelegramBotAPI.types.typeField
+import com.github.insanusmokrassar.TelegramBotAPI.utils.nonstrictJsonFormat
 import kotlinx.serialization.*
+import kotlinx.serialization.json.JsonObjectSerializer
 
 @Serializer(MediaGroupMemberInputMedia::class)
 internal object MediaGroupMemberInputMediaSerializer : KSerializer<MediaGroupMemberInputMedia> {
@@ -13,6 +16,12 @@ internal object MediaGroupMemberInputMediaSerializer : KSerializer<MediaGroupMem
     }
 
     override fun deserialize(decoder: Decoder): MediaGroupMemberInputMedia {
-        TODO("not implemented")
+        val json = JsonObjectSerializer.deserialize(decoder)
+
+        return when (json.getPrimitiveOrNull(typeField) ?.contentOrNull) {
+            photoInputMediaType -> nonstrictJsonFormat.fromJson(InputMediaPhoto.serializer(), json)
+            videoInputMediaType -> nonstrictJsonFormat.fromJson(InputMediaVideo.serializer(), json)
+            else -> error("Illegal type of incoming MediaGroupMemberInputMedia")
+        }
     }
 }