+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/JSON.kt
@@ -3,6 +3,7 @@ package com.github.insanusmokrassar.TelegramBotAPI.utils
 import kotlinx.serialization.SerializationStrategy
 import kotlinx.serialization.json.*
 
+@Suppress("EXPERIMENTAL_API_USAGE")
 internal val nonstrictJsonFormat = Json {
     isLenient = true
     ignoreUnknownKeys = true